+++ render/cw1/shaders/point.vert
@@ -4,7 +4,7 @@
 
 layout(location = 0) in vec3 iPosition;
 layout(location = 1) in vec3 iColor;
-layout(location = 2) in float iScalar;
+layout(location = 2) in int iScalar;
 
 layout( set = 0, binding = 0) uniform UScene{
     mat4 camera;
